+++ src/pages/noticedetail/component/CustomModal/CustomModal.styled.ts
