+++ src/pages/noticedetail/component/CustomModal/CustomModal.tsx
