+++ src/pages/noticedetail/component/NoticePostList/NoticePostList.styled.ts
